test(unit): move backend health tests to match new structure
